// packages/sdk-typescript/src/generated/api/events-api.ts
export interface EventsApiInterface {
    /**
     * 
     * @summary Create new event
     * @param {CreateEventRequest} createEventRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApiInterface
     */
    createEvent(createEventRequest: CreateEventRequest, options?: RawAxiosRequestConfig): AxiosPromise<StandardResponse>;

    /**
     * 
     * @summary Delete event
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApiInterface
     */
    deleteEvent(id: string, options?: RawAxiosRequestConfig): AxiosPromise<void>;

    /**
     * 
     * @summary Get event details
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApiInterface
     */
    getEventById(id: string, options?: RawAxiosRequestConfig): AxiosPromise<StandardResponse>;

    /**
     * 
     * @summary List all events
     * @param {number} [page] 
     * @param {number} [limit] 
     * @param {GetEventsStatusEnum} [status] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApiInterface
     */
    getEvents(page?: number, limit?: number, status?: GetEventsStatusEnum, options?: RawAxiosRequestConfig): AxiosPromise<PaginatedResponse>;

    /**
     * 
     * @summary Update event
     * @param {string} id 
     * @param {Event} event 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApiInterface
     */
    updateEvent(id: string, event: Event, options?: RawAxiosRequestConfig): AxiosPromise<void>;

}

/**
 * EventsApi - object-oriented interface
 * @export
 * @class EventsApi
 * @extends {BaseAPI}
 */
export class EventsApi extends BaseAPI implements EventsApiInterface {
    /**
     * 
     * @summary Create new event
     * @param {CreateEventRequest} createEventRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApi
     */
    public createEvent(createEventRequest: CreateEventRequest, options?: RawAxiosRequestConfig) {
        return EventsApiFp(this.configuration).createEvent(createEventRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Delete event
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApi
     */
    public deleteEvent(id: string, options?: RawAxiosRequestConfig) {
        return EventsApiFp(this.configuration).deleteEvent(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get event details
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApi
     */
    public getEventById(id: string, options?: RawAxiosRequestConfig) {
        return EventsApiFp(this.configuration).getEventById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary List all events
     * @param {number} [page] 
     * @param {number} [limit] 
     * @param {GetEventsStatusEnum} [status] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApi
     */
    public getEvents(page?: number, limit?: number, status?: GetEventsStatusEnum, options?: RawAxiosRequestConfig) {
        return EventsApiFp(this.configuration).getEvents(page, limit, status, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update event
     * @param {string} id 
     * @param {Event} event 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EventsApi
     */
    public updateEvent(id: string, event: Event, options?: RawAxiosRequestConfig) {
        return EventsApiFp(this.configuration).updateEvent(id, event, options).then((request) => request(this.axios, this.basePath));
    }
}

/**
 * @export
 */
export const GetEventsStatusEnum = {
    Draft: 'draft',
    Published: 'published',
    SoldOut: 'sold_out',
    Cancelled: 'cancelled',
    Completed: 'completed'
} as const;
export type GetEventsStatusEnum = typeof GetEventsStatusEnum[keyof typeof GetEventsStatusEnum];